Provide two transforms: `attachment` and `attachments`
