+++ app/transforms/attachments.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-pouch/transforms/attachments';
